Added an admin boolean to the user class
I added an admin boolean (isAdmin) that allows you to know if the user
is an admin.
      - True means they are an admin
      - False means they aren't an admin
I assumed that we wouldn't set admin on registration, becasue that seems
like too much power for the user, so I also included two methods,
setAsAdmin and removeAsAdmin, to allow the power to be added or removed
from the user. I also added a getter to the class for the isAdmin
boolean so you can later tell if they are an admin or not.